front-end/router: bug with refactoring imported contentwrapper incorrectly
