+++ src/front-end/state/setupDailyLog.js
@@ -1,10 +1,10 @@
 import * as localStorage from "../localStorage/userOperations.js";
-import { contentWrapper, setPageType, setUrl } from "./router.js";
+import { setPageType, setUrl } from "./router.js";
 import { TrackerBlock } from "../components/trackerBlock.jsx";
 import { TrackerMenu } from "../components/tracker.jsx";
 import { createEditor } from "../components/blockController.js";
 import { currentState } from "./stateManager.js";
-import { header } from "../index.js";
+import { header, contentWrapper } from "../index.js";
 
 
 /**
