+++ src/front-end/state/setupIndex.js
@@ -1,9 +1,9 @@
 import * as localStorage from "../localStorage/userOperations.js";
-import { contentWrapper, setPageType, setUrl } from "./router.js";
+import { setPageType, setUrl } from "./router.js";
 import { CreatorBlock } from "../components/creator.jsx";
 import { DropdownBlock } from "../components/dropdown.jsx";
 import { currentState } from "./stateManager.js";
-import { header } from "../index.js"
+import { header, contentWrapper } from "../index.js"
 
 
 /**
